fix: adjusting to the new API from eilos 1.0.0-rc15
