+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "eilos-preset-typescript",
-  "version": "1.0.0-rc4",
+  "version": "1.0.0-rc5",
   "description": "Web App Build Abstraction",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -65,7 +65,7 @@
     "@types/copy-webpack-plugin": "^8.0.1",
     "@types/mini-css-extract-plugin": "^2.2.0",
     "ajv": "^8.6.2",
-    "eilos": "^1.0.0-rc9",
+    "eilos": "1.0.0-rc15",
     "ts-loader": "^9.2.5",
     "typescript": "^4.3.5",
     "webpack": "^5.50.0",
