--- src/config/eslint.config.ts
@@ -1,6 +1,0 @@
-import type { ConfigFileContents } from "eilos";
-import type { PresetRuntimeContext } from "../options";
-
-export default function (ctx: PresetRuntimeContext): ConfigFileContents {
-  return {};
-}
