+++ src/files/prettier.config.ts
@@ -1,0 +1,25 @@
+import { DefinePresetFile } from "eilos";
+
+import { Options } from "../options";
+
+const file = DefinePresetFile(Options, {
+  mimeType: "application/javascript",
+  generator: (ctx) => {
+    const { merge } = ctx.util;
+    const userConfig = ctx.getConfig("prettier", {});
+
+    // Base 'prettier.config.js' contents
+    const BaseConfig = {
+      trailingComma: "es5",
+      tabWidth: 2,
+      semi: true,
+      singleQuote: false,
+      printWidth: 100,
+      useTabs: false,
+    };
+
+    return merge(BaseConfig, userConfig);
+  },
+});
+
+export default file;
